Add files via upload
When I tried to split the text so that it is only the beginning to end, it would not print properly in the console so I had to split it up for it to print the entire story. Im still a little confused on what to print out in the draw function. I tried manipulating the example from class but that didn't do anything.